Fixing equalType to stop using pointer equality
- Stop using pointer equality with NUM, BOOL and NONE constants in
  equalType. This makes things error prone and brittle.
- Use equalType to determine the type of the argument to the print
  builtin function.
- Adding back the type-checking for the argument to print using
  equalType and also allowing the argument to be NONE.

The above two changes fix the print built in not working in Generic
classes. This was happening because the Monomorphizer makes a deep copy
of the entire AST of the class when concretizing it which does not
preserve the pointer addresses of primitive type annotations like NUM,
BOOL and NONE. This caused the compiler to be unable to determine the
type of the argument to print.